pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.1.0",
    "huggingface-hub>=0.28.1",
    "openai>=1.62.0",
    "python-dotenv>=1.0.1",
]
